Implementações

+++ src/main/java/br/edu/utfpr/dungeontable/repository/MagicRepository.java
@@ -3,5 +3,5 @@ package br.edu.utfpr.dungeontable.repository;
 import br.edu.utfpr.dungeontable.model.tools.Magic;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MagicRepository extends JpaRepository<Magic,Integer> {
+public interface MagicRepository extends JpaRepository<Magic,Long> {
 }
